[data] fix array contain nan element in group by Signed-off-by: jukejian <jukejian@bytedance.com>
